+++ src/state-builtin.ts
@@ -1,0 +1,10 @@
+import {RngState} from "./state";
+
+export class RngStateBuiltin extends RngState {
+    constructor() {
+        super();
+    }
+    public random() {
+        return Math.random();
+    }
+}
